requirements.txt
fastapi
uvicorn
jinja2
pydantic
paho-mqtt
RPi.GPIO
pytest
python-dotenv
